requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
aiofiles==24.1.0
fastapi==0.115.3
uvicorn==0.32.0
aiohttp==3.10.6
openai==1.49.0
pydub==0.25.1